+++ docs/CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+* Add translations support, with english and spanish languages
+* Add "exported on" timestamp on importing view
 
 
 ## 1.10.1 - 2019-11-09